+++ examples/js/cli/helpers.js
@@ -1,4 +1,4 @@
-const listAll = async (connection) => {
+const listCharacters = async (connection) => {
   const characters = await connection.getCharacters();
 
   if (characters.length) {
@@ -27,8 +27,50 @@ const changeCharacter = async (connection, id) => {
   }
 };
 
+const conversationInfo = async (conversation) => {
+  console.log(`ConversartionId: ${conversation.getConversationId()}`);
+  conversation.getCharacters().forEach((c) => {
+    console.log(`${conversation.getConversationId()}: ${characterInfo(c)}`);
+  });
+  console.log('------------------------------');
+};
+
+const listConversations = async (connection) => {
+  const conversations = connection.getConversations();
+
+  if (conversations.length) {
+    conversations.map((c) => {
+      console.log(`Conversartion (${c.conversationId})`);
+      c.characters.forEach((c) => {
+        console.log(characterInfo(c));
+      });
+      console.log('------------------------------');
+    });
+  } else {
+    console.log('No conversations');
+  }
+};
+
+const changeScene = async (connection, scene) => {
+  try {
+    await connection.changeScene(scene);
+    console.log('Scene changed');
+  } catch (e) {
+    console.log('Scene not changed: ', e.message);
+  }
+};
+
+const addCharacters = async (connection, characters) => {
+  const done = await connection.addCharacters(characters);
+  console.log(done ? 'Characters loaded' : 'Characters not loaded.');
+};
+
 module.exports = {
-  listAll,
+  addCharacters,
   changeCharacter,
+  changeScene,
   characterInfo,
+  conversationInfo,
+  listCharacters,
+  listConversations,
 };