+++ proto/ai/inworld/engine/v1/state_serialization_pb.d.ts
@@ -41,6 +41,11 @@ export class SessionState extends jspb.Message {
     getCreationTime(): google_protobuf_timestamp_pb.Timestamp | undefined;
     setCreationTime(value?: google_protobuf_timestamp_pb.Timestamp): SessionState;
 
+    hasVersion(): boolean;
+    clearVersion(): void;
+    getVersion(): SessionState.Version | undefined;
+    setVersion(value?: SessionState.Version): SessionState;
+
     serializeBinary(): Uint8Array;
     toObject(includeInstance?: boolean): SessionState.AsObject;
     static toObject(includeInstance: boolean, msg: SessionState): SessionState.AsObject;
@@ -55,5 +60,28 @@ export namespace SessionState {
     export type AsObject = {
         state: Uint8Array | string,
         creationTime?: google_protobuf_timestamp_pb.Timestamp.AsObject,
+        version?: SessionState.Version.AsObject,
     }
+
+
+    export class Version extends jspb.Message { 
+        getInteractionId(): string;
+        setInteractionId(value: string): Version;
+
+        serializeBinary(): Uint8Array;
+        toObject(includeInstance?: boolean): Version.AsObject;
+        static toObject(includeInstance: boolean, msg: Version): Version.AsObject;
+        static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
+        static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
+        static serializeBinaryToWriter(message: Version, writer: jspb.BinaryWriter): void;
+        static deserializeBinary(bytes: Uint8Array): Version;
+        static deserializeBinaryFromReader(message: Version, reader: jspb.BinaryReader): Version;
+    }
+
+    export namespace Version {
+        export type AsObject = {
+            interactionId: string,
+        }
+    }
+
 }